+++ extension_packages/figpack_nwb/figpack_nwb/views/__init__.py
@@ -1,3 +1,4 @@
 from .PlaneSegmentation import PlaneSegmentation
+from .PoseEstimation import PoseEstimation, PoseEstimationItem
 
-__all__ = ["PlaneSegmentation"]
+__all__ = ["PlaneSegmentation", "PoseEstimation", "PoseEstimationItem"]